% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorables.R
\name{copy_explorables}
\alias{copy_explorables}
\title{Copy explorables to www directory}
\usage{
copy_explorables(url)
}
\arguments{
\item{url}{url to the zipfile containig explorables (explorables.zip)}
}
\description{
Copy explorables to www directory
}
