% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorables.R
\name{create_explorable}
\alias{create_explorable}
\title{Create explorable}
\usage{
create_explorable(path_dir)
}
\arguments{
\item{path_dir}{Path to the directory containing the explorable. The Rmd is
expected to be named index.Rmd.}
}
\description{
Create explorable
}
