% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_fluid_page.R
\name{dc_fluid_page}
\alias{dc_fluid_page}
\title{Create a conceptual exercise with fluid layout}
\usage{
dc_fluid_page(..., title = NULL, theme = NULL,
  local = Sys.info()[["sysname"]] == "Darwin")
}
\description{
Create a conceptual exercise with fluid layout
}
\examples{
num <- sample(1:10, 1)
ui <- dc_fluid_page(
  theme = shinythemes::shinytheme('cosmo'),
  titlePanel('Guess the Number'),
  mainPanel(
    sliderInput('num', 'Select number', 0, 10, 2),
    actionButton('submit', label = 'Submit'),
    uiOutput('num')
  )
)
server <- function(input, output, session){
   payload <- reactive({
     num_selected <- input$num
     message <- if (num_selected == num){
       list(success = TRUE, message = 'Well Done!')
     } else {
       list(success = FALSE, message = 'Sorry! Try Again...')
     }
   })
   observeEvent(input$submit, {
     session$sendCustomMessage("campus", payload())
   })
 }
 if (interactive()){
   shinyApp(ui = ui, server = server)
}

dc_fluid_page(
  titlePanel('Guess the Number'),
  mainPanel(
    tags$input("v-model" = "num",
      type = 'number', min = "0", max = "10", step = "1"
    ),
    tags$button(type = "submit", "Submit Answer", `@click` = "submit")
  ) \%>\%
  vuepkg::vue(
    data = list(num = 2),
    methods = list(
      submit = htmlwidgets::JS("function(){
         if (this.num == 6){
           checkExercise(true, 'Well done!')
         } else {
           checkExercise(false, 'Sorry, Try Again!')
         }
      }")
    )
  )
) \%>\%
htmltools::browsable()
}
