% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_parsons.R
\name{dc_parsons-shiny}
\alias{dc_parsons-shiny}
\alias{dc_parsonsOutput}
\alias{renderDc_parsons}
\title{Shiny bindings for dc_parsons}
\usage{
dc_parsonsOutput(outputId, width = "100\%", height = "400px")

renderDc_parsons(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a dc_parsons}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using dc_parsons within Shiny
applications and interactive Rmd documents.
}
