% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_parsons.R
\name{dc_parsons}
\alias{dc_parsons}
\title{Create a Parson's Problem Exercise}
\usage{
dc_parsons(code, campus = FALSE, width = "100\%", height = "auto",
  elementId = NULL)
}
\description{
Create a Parson's Problem Exercise
}
\examples{
dc_parsons("def is_true(boolean_value):\\n  if boolean_value:\\n     return True\\n  return False\\n
 return true #distractor\\n"
)
}
