% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_www_dir}
\alias{get_www_dir}
\title{Get path to www directory}
\usage{
get_www_dir(local_www_dir = tempdir())
}
\arguments{
\item{local_www_dir}{local www directory}
}
\description{
Get path to www directory
}
