% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noncoding.R
\name{non_coding}
\alias{non_coding}
\alias{non_coding_ui}
\title{Shiny module to create a non-coding exercise}
\usage{
non_coding(input, output, session, feedback)

non_coding_ui(id)
}
\arguments{
\item{input}{standard \code{shiny} boilerplate}

\item{output}{standard \code{shiny} boilerplate}

\item{session}{standard \code{shiny} boilerplate}

\item{feedback}{a reactive that returns a named list with a
message and feedback}

\item{id}{id to call the module with}
}
\description{
Shiny module to create a non-coding exercise
}
\examples{
\dontrun{
 library(shiny)
 num <- sample(1:10, 1)

 check_exercise <- function(actual, guess){
   if(actual == guess){
     list(message = "Well Done!", success = TRUE)
   } else {
     msg <- sprintf("You guessed \%s. Please try again!", guess)
     list(message = msg, success = FALSE)
   }
 }

 ui <- fluidPage(
   theme = shinythemes::shinytheme('cosmo'),
   titlePanel('Guess the Number'),
   mainPanel(
     sliderInput('num', 'Select number', 0, 10, 2),
     non_coding_ui('feedback')
   )
 )

 server <- function(input, output, session){
   feedback <- reactive({
     check_exercise(num, input$num)
   })
   shiny::callModule(non_coding, 'feedback', feedback)
 }
 shinyApp(ui = ui, server = server)
}
}
