% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_fun}
\alias{write_fun}
\title{Write function to Rprofile.site}
\usage{
write_fun(fun, file = "/etc/R/Rprofile.site")
}
\arguments{
\item{fun}{function to write to Rprofile.site}

\item{file}{the file to write the function to}
}
\description{
This will make the function globally available to all exercises.
}
\examples{
\dontrun{
 tf <- tempfile(fileext = '.R')
 add <- function(x, y){
   x + y
 }
 write_fun(add, file = tf)
 cat(paste(readLines(tf, warn = FALSE), collapse = '\\n'))
}
}
