% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorables.R
\name{zip_explorables}
\alias{zip_explorables}
\title{Zip a folder containing explorables}
\usage{
zip_explorables(path = "explorables")
}
\arguments{
\item{path}{path to explorables}
}
\description{
Zip a folder containing explorables
}
