% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorables.R
\name{display_explorable}
\alias{display_explorable}
\title{Display Explorable in the HTML Viewer}
\usage{
display_explorable(path, subdir = Sys.getenv("DC_EXPLORABLES_SUBDIR",
  ""), cdn = Sys.getenv("DC_EXPLORABLES_CDN", FALSE),
  type = Sys.getenv("DC_EXPLORABLES_TYPE", "dev"), version = "latest",
  index = "index.html")
}
\arguments{
\item{path}{Path to the folder containing index.html.}

\item{subdir}{The parent directory for the course.}

\item{cdn}{A boolean indicating if asset should be served from CDN.}

\item{type}{A string indicating whether to use dev, staging, or prod.}

\item{version}{A string indicating the version of the asset.}

\item{index}{Name of the html file (defaults to index.html)}
}
\description{
This function can be used to display an explorable web app in a
}
